pyproject.toml
[project]
name = "TheOpenDataStack"
version = "0.1.0"
description = "Build end to end data products with open source tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dagster==1.10.2",    
    "dagster-dbt==0.26.2",
    "dagster-duckdb>=0.26.2",
    "dagster-webserver==1.10.2",
    "dbt-core>=1.8.7",
    "dbt-duckdb>=1.8.3",
    "duckdb==1.0.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "polars>=1.8.0",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.1",
    "python-slugify>=8.0.4",
    "requests>=2.32.3",
    "tenacity>=9.0.0",
    "ruff>=0.6.7",
    "httpx<0.25.0",
    "flash>=1.0.3",
    "flask>=3.1.0",
    "great-tables>=0.16.1",
]

[tool.dagster]
module_name = "pipeline.definitions"
code_location_name = "pipeline"